+++ changelog.md
@@ -2,6 +2,10 @@
 
 ## Release 0.18.12
 
+### Features
+
+- Automatically update Swing, Thrust, and Basic Lift when updating ST.
+
 ### Bugfixes
 
 - Modern UI sheet uses `GURPS.LastActor` as actor target instead of sheet actor \#2472
@@ -19,6 +23,7 @@
 - Restore Foundry v12 compatibility \#2500
 - ADD tool tips missing \#2503
 - Fix blank usermods being created.
+- Error when player chooses to end his turn in combat \#2502
 
 ## Release 0.18.11 01/09/2026